Extract inserter state into reusable hooks
